Two TOFS working on Teensy 4.0
Changed the code based on an example found, adapted for this use. Two sensors are hooked up and in use with this code and running on Teensy 4.0. Next step will be getting it to run on the Teensy 4.1 when they arrive, chaining in the third TOFS, then hooking up lights to be triggered by the TOFS.